Remove redux and refactor types
