+++ client/package.json
@@ -12,14 +12,12 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
-    "@types/react-redux": "^7.1.15",
     "@types/react-router-dom": "^5.1.7",
     "@types/socket.io-client": "^1.4.34",
     "@types/styled-components": "^5.1.4",
     "axios": "^0.21.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
     "socket.io-client": "^3.0.5",
