+++ client/src/actions/game.ts
@@ -1,13 +1,5 @@
 import api from "utils/api";
 
-export interface Game {
-  name: string;
-  type: number;
-  code: string;
-  users: { id: string, name: string; hand: number | null }[];
-  revealed: boolean;
-}
-
 export class GameNotFound extends Error {
   code: number;
   constructor(message: string | undefined) {
