--- client/src/store/index.ts
@@ -1,8 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import rootReducer from "./reducers";
-
-const store = configureStore({
-  reducer: rootReducer,
-});
-
-export default store
