--- client/src/store/reducers/index.ts
@@ -1,11 +1,0 @@
-import { combineReducers } from '@reduxjs/toolkit'
-
-import game from './game'
-
-const rootReducer = combineReducers({
-  game
-})
-
-export type RootState = ReturnType<typeof rootReducer>
-
-export default rootReducer
