+++ client/src/types/Profile.ts
@@ -1,0 +1,5 @@
+export interface Profile {
+  name: string;
+  id: string;
+  hand?: number;
+}
